// src/api/endpoints/api.schemas.ts
export interface RoomPerms {
  roomId?: string;
  /** @nullable */
  userId?: string | null;
  renovationOnly?: boolean;
  canEdit?: boolean;
}